BAQE-586: Fix issues revealed by SonarQube (#2153)
* BAQE-536: Fix possible dereference of missing Optional

* BAQE-586: Fix double bracket initialization

This PR fixes the issue revealed by findbugs:
"Because Double Brace Initialization (DBI) creates an anonymous class with a reference to the instance of the owning object, its use can lead to memory leaks if the anonymous inner class is returned and held by other objects. Even when there's no leak, DBI is so obscure that it's bound to confuse most maintainers."

* BAQE-586: Fix code formatting following code review

+++ kie-wb-common-dmn/kie-wb-common-dmn-client/src/main/java/org/kie/workbench/common/dmn/client/editors/expressions/types/dtable/DecisionTableDefaultValueUtilities.java
@@ -40,7 +40,7 @@ public class DecisionTableDefaultValueUtilities {
         return INPUT_CLAUSE_PREFIX + getMaxUnusedInputClauseIndex(dtable);
     }
 
-    private static long getMaxUnusedInputClauseIndex(final DecisionTable dtable) {
+    private static int getMaxUnusedInputClauseIndex(final DecisionTable dtable) {
         int maxIndex = 0;
         for (InputClause ic : dtable.getInput()) {
             final LiteralExpression le = ic.getInputExpression();
@@ -58,7 +58,7 @@ public class DecisionTableDefaultValueUtilities {
         return OUTPUT_CLAUSE_PREFIX + getMaxUnusedOutputClauseIndex(dtable);
     }
 
-    private static long getMaxUnusedOutputClauseIndex(final DecisionTable dtable) {
+    private static int getMaxUnusedOutputClauseIndex(final DecisionTable dtable) {
         int maxIndex = 0;
         for (OutputClause oc : dtable.getOutput()) {
             final Optional<Integer> index = DefaultValueUtilities.extractIndex(oc.getName(), OUTPUT_CLAUSE_PREFIX);
