+++ kie-wb-common-dmn/kie-wb-common-dmn-client/src/main/java/org/kie/workbench/common/dmn/client/property/dmn/DefaultValueUtilities.java
@@ -70,8 +70,8 @@ public class DefaultValueUtilities {
         }
     }
 
-    public static long getMaxUnusedIndex(final Collection<String> values,
-                                         final String prefix) {
+    public static int getMaxUnusedIndex(final Collection<String> values,
+                                        final String prefix) {
         int maxIndex = 0;
         for (final String value : values) {
             final Optional<Integer> index = extractIndex(value, prefix);
